first impressions on soundscape
After joining the project and learning about it, I started keeping little notes about my process. I was recommended to do so by Professor Turner for new students that might join the project later on. This is the first of several .txt files I made taking notes on the process/my work.